Make frame_control part of header (remove dublicated AddressMode information); Add header module
