+++ src/mac/frame/mod.rs
@@ -12,13 +12,14 @@
 // - remove one variant enums
 
 use bytes::{Buf, BufMut};
-use hash32_derive::Hash32;
 
 use crate::mac::beacon::Beacon;
 use crate::mac::command::Command;
 
-pub mod frame_control;
-pub use frame_control::{AddressMode, FrameControl, FrameType};
+mod frame_control;
+pub mod header;
+use header::FrameType;
+pub use header::Header;
 
 /// An IEEE 802.15.4 MAC frame
 ///
@@ -61,15 +62,17 @@ impl<'p> Frame<'p> {
     /// # Example
     ///
     /// ``` rust
-    /// use ieee802154::mac::{
-    ///     Address,
-    ///     ShortAddress,
+    /// use ieee802154::mac::frame::{
     ///     Frame,
-    ///     FrameType,
-    ///     PanId,
-    /// };
+    ///     header::{
+    ///       Address,
+    ///       ShortAddress,
+    ///       FrameType,
+    ///       PanId,
+    ///       Security
+    /// }};
     ///
-    /// # fn main() -> Result<(), ::ieee802154::mac::DecodeError> {
+    /// # fn main() -> Result<(), ::ieee802154::mac::frame::DecodeError> {
     /// // Construct a simple MAC frame. The CRC checksum (the last 2 bytes) is
     /// // invalid, for the sake of convenience.
     /// let bytes = [
@@ -82,14 +85,14 @@ impl<'p> Frame<'p> {
     /// ];
     ///
     /// let frame = Frame::decode(&bytes, true)?;
-    /// let fcnt = frame.header.frame_control;
+    /// let header = frame.header;
     ///
-    /// assert_eq!(frame.header.seq,             0x00);
-    /// assert_eq!(fcnt.frame_type,      FrameType::Data);
-    /// assert_eq!(fcnt.security,        false);
-    /// assert_eq!(fcnt.frame_pending,   false);
-    /// assert_eq!(fcnt.ack_request,     false);
-    /// assert_eq!(fcnt.pan_id_compress, false);
+    /// assert_eq!(frame.header.seq,       0x00);
+    /// assert_eq!(header.frame_type,      FrameType::Data);
+    /// assert_eq!(header.security,        Security::None);
+    /// assert_eq!(header.frame_pending,   false);
+    /// assert_eq!(header.ack_request,     false);
+    /// assert_eq!(header.pan_id_compress, false);
     ///
     /// assert_eq!(
     ///     frame.header.destination,
@@ -148,34 +151,30 @@ impl<'p> Frame<'p> {
     /// # Example
     ///
     /// ``` rust
-    /// use ieee802154::mac::{
+    /// use ieee802154::mac::frame::{
+    ///   Frame,
+    ///   FrameContent,
+    ///   WriteFooter,
+    ///   header::{
     ///     Address,
     ///     ShortAddress,
-    ///     Frame,
-    ///     FrameContent,
     ///     FrameType,
     ///     FrameVersion,
     ///     Header,
     ///     PanId,
-    ///     WriteFooter,
-    ///     FrameControl,
-    ///     AddressMode,
-    /// };
+    ///     Security,
+    /// }};
     ///
     /// let frame = Frame {
     ///     header: Header {
-    ///         frame_control: FrameControl {
     ///         frame_type:      FrameType::Data,
-    ///         security:        false,
+    ///         security:        Security::None,
     ///         frame_pending:   false,
     ///         ack_request:     false,
     ///         pan_id_compress: false,
     ///         version:         FrameVersion::Ieee802154_2006,
-    ///         dest_addr_mode: AddressMode::Short,
-    ///         src_addr_mode: AddressMode::Short,
-    ///         },
-    ///         seq:             0x00,
     ///
+    ///         seq:             0x00,
     ///         destination: Some(Address::Short(PanId(0x1234), ShortAddress(0x5678))),
     ///         source:      Some(Address::Short(PanId(0x1234), ShortAddress(0x9abc))),
     ///     },
@@ -229,478 +228,6 @@ pub enum WriteFooter {
     No,
 }
 
-/// MAC frame header
-///
-/// External documentation for [MAC frame format start at 5.2]
-///
-/// [MAC frame format start at 5.2]: http://ecee.colorado.edu/~liue/teaching/comm_standards/2015S_zigbee/802.15.4-2011.pdf
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub struct Header {
-    /// Frame Control Field
-    pub frame_control: FrameControl,
-
-    /// Sequence Number
-    pub seq: u8,
-
-    /// Destination Address
-    pub destination: Option<Address>,
-
-    /// Source Address
-    pub source: Option<Address>,
-}
-
-impl Header {
-    /// Decodes a header from a byte buffer
-    ///
-    /// This method is used by [`Frame::decode`] to decode the frame header.
-    /// Unless you decide to write your own code for decoding frames, there
-    /// should be no reason to call this method directly.
-    ///
-    /// # Errors
-    ///
-    /// This function returns an error, if the bytes either don't encode a valid
-    /// IEEE 802.15.4 frame header, or encode a frame header that is not fully
-    /// supported by this implementation. Please refer to [`DecodeError`] for
-    /// details.
-    ///
-    /// # Example
-    ///
-    /// ``` rust
-    /// use ieee802154::mac::{
-    ///     Address,
-    ///     ShortAddress,
-    ///     FrameType,
-    ///     Header,
-    ///     PanId,
-    /// };
-    ///
-    /// # fn main() -> Result<(), ::ieee802154::mac::DecodeError> {
-    /// // Construct a simple header.
-    /// let mut bytes = &[
-    ///     0x01, 0x98,             // frame control
-    ///     0x00,                   // sequence number
-    ///     0x12, 0x34, 0x56, 0x78, // PAN identifier and address of destination
-    ///     0x12, 0x34, 0x9a, 0xbc, // PAN identifier and address of source
-    /// ][..];
-    ///
-    /// let header = Header::decode(&mut bytes)?;
-    /// let fctl = header.frame_control;
-    ///
-    /// assert_eq!(fctl.frame_type,      FrameType::Data);
-    /// assert_eq!(fctl.security,        false);
-    /// assert_eq!(fctl.frame_pending,   false);
-    /// assert_eq!(fctl.ack_request,     false);
-    /// assert_eq!(fctl.pan_id_compress, false);
-    /// assert_eq!(header.seq,             0x00);
-    ///
-    /// assert_eq!(
-    ///     header.destination,
-    ///     Some(Address::Short(PanId(0x3412), ShortAddress(0x7856)))
-    /// );
-    /// assert_eq!(
-    ///     header.source,
-    ///     Some(Address::Short(PanId(0x3412), ShortAddress(0xbc9a)))
-    /// );
-    /// #
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn decode(buf: &mut dyn Buf) -> Result<Self, DecodeError> {
-        // First, make sure we have enough buffer for the Frame Control field
-        if buf.remaining() < 3 {
-            return Err(DecodeError::NotEnoughBytes);
-        }
-
-        let fctl = FrameControl::try_from_bits(buf.get_u16_le())?;
-
-        let seq = buf.get_u8();
-
-        let destination = Address::decode(buf, &fctl.dest_addr_mode)?;
-
-        let source = if !fctl.pan_id_compress {
-            Address::decode(buf, &fctl.src_addr_mode)?
-        } else {
-            let pan_id = destination
-                .ok_or(DecodeError::InvalidAddressMode(fctl.dest_addr_mode.as_u8()))?
-                .pan_id();
-            Address::decode_compress(buf, &fctl.src_addr_mode, pan_id)?
-        };
-
-        let header = Header {
-            frame_control: fctl,
-            seq,
-            destination,
-            source,
-        };
-
-        Ok(header)
-    }
-
-    /// Encodes the header into a buffer
-    ///
-    /// The header length depends on the options chosen and varies between 3 and
-    /// 30 octets.
-    ///
-    /// # Example
-    ///
-    /// ``` rust
-    /// use bytes::BytesMut;
-    /// use ieee802154::mac::{
-    ///     Address,
-    ///     AddressMode,
-    ///     ShortAddress,
-    ///     FrameType,
-    ///     FrameVersion,
-    ///     Header,
-    ///     PanId,
-    ///     FrameControl,
-    /// };
-    ///
-    /// let header = Header {
-    ///     frame_control: FrameControl {
-    ///     frame_type:      FrameType::Data,
-    ///     security:        false,
-    ///     frame_pending:   false,
-    ///     ack_request:     false,
-    ///     pan_id_compress: false,
-    ///     version:         FrameVersion::Ieee802154_2006,
-    ///         dest_addr_mode: AddressMode::Short,
-    ///         src_addr_mode: AddressMode::Short,
-    ///     },
-    ///     seq:             0x00,
-    ///
-    ///     destination: Some(Address::Short(PanId(0x1234), ShortAddress(0x5678))),
-    ///     source:      Some(Address::Short(PanId(0x1234), ShortAddress(0x9abc))),
-    /// };
-    ///
-    /// let mut bytes = BytesMut::with_capacity(11);
-    ///
-    /// header.encode(&mut bytes);
-    /// let encoded_bytes = bytes.split().freeze();
-    ///
-    /// let expected_bytes = [
-    ///     0x01, 0x98,             // frame control
-    ///     0x00,                   // sequence number
-    ///     0x34, 0x12, 0x78, 0x56, // PAN identifier and address of destination
-    ///     0x34, 0x12, 0xbc, 0x9a, // PAN identifier and address of source
-    /// ];
-    /// assert_eq!(encoded_bytes, expected_bytes[..]);
-    /// ```
-    pub fn encode(&self, buf: &mut dyn BufMut) {
-        let frame_control_raw = self.frame_control.to_bits();
-
-        buf.put_u16_le(frame_control_raw);
-
-        // Write Sequence Number
-        buf.put_u8(self.seq);
-
-        // Write addresses
-        if let Some(destination) = self.destination {
-            assert!(self.frame_control.dest_addr_mode != AddressMode::None);
-            destination.encode(buf);
-        }
-
-        match (self.source, self.frame_control.pan_id_compress) {
-            (Some(source), true) => {
-                assert!(self.frame_control.src_addr_mode != AddressMode::None);
-                source.encode_compress(buf);
-            }
-            (Some(source), false) => {
-                assert!(self.frame_control.src_addr_mode != AddressMode::None);
-                source.encode(buf);
-            }
-            (None, true) => panic!("Frame cnt request commpress address without address"),
-            (None, false) => (),
-        }
-    }
-}
-
-/// Personal Area Network Identifier
-///
-/// A 16-bit value that identifies a PAN
-///
-/// # Example
-///
-/// ``` rust
-/// use ieee802154::mac::frame::PanId;
-///
-/// let pan_id = PanId(0x0123);
-/// ```
-#[derive(Clone, Copy, Debug, Eq, Hash, Hash32, PartialEq)]
-pub struct PanId(pub u16);
-
-impl PanId {
-    /// Get the broadcast PAN identifier
-    pub fn broadcast() -> Self {
-        Self(0xffff)
-    }
-
-    /// Decodes an PAN identifier from a byte buffer
-    ///
-    /// # Errors
-    ///
-    /// This function returns an error, if there are not enough bytes in the
-    /// buffer to encode a valid `Address` instance.
-    ///
-    /// # Example
-    ///
-    /// ``` rust
-    /// use ieee802154::mac::PanId;
-    ///
-    /// # fn main() -> Result<(), ::ieee802154::mac::DecodeError> {
-    /// let mut bytes = &[0x56, 0x78][..];
-    /// let address = PanId::decode(&mut bytes)?;
-    ///
-    /// assert_eq!(address.0, 0x7856);
-    /// #
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn decode(buf: &mut dyn Buf) -> Result<Self, DecodeError> {
-        if buf.remaining() < 2 {
-            return Err(DecodeError::NotEnoughBytes);
-        }
-
-        Ok(PanId(buf.get_u16_le()))
-    }
-
-    /// Encodes the PAN identifier into a buffer
-    ///
-    /// # Example
-    ///
-    /// ``` rust
-    /// use ieee802154::mac::PanId;
-    /// use bytes::BytesMut;
-    ///
-    /// let address = PanId(0x1234);
-    ///
-    /// let mut bytes = BytesMut::with_capacity(2);
-    /// address.encode(&mut bytes);
-    ///
-    /// let expected_bytes = [0x34, 0x12];
-    /// assert_eq!(bytes[..], expected_bytes[..]);
-    /// ```
-    pub fn encode(&self, buf: &mut dyn BufMut) {
-        buf.put_u16_le(self.0)
-    }
-}
-
-/// A 16-bit short address
-///
-/// Short address assigned to a device during association, used to identify the
-/// device in the PAN.
-///
-/// # Example
-///
-/// ``` rust
-/// use ieee802154::mac::frame::ShortAddress;
-///
-/// let short_address = ShortAddress(0x0123);
-/// ```
-#[derive(Clone, Copy, Debug, Eq, Hash, Hash32, PartialEq)]
-pub struct ShortAddress(pub u16);
-
-impl ShortAddress {
-    /// An instance of `ShortAddress` that represents the broadcast address.
-    pub const BROADCAST: Self = ShortAddress(0xffff);
-
-    /// Creates an instance of `ShortAddress` that represents the broadcast address
-    pub fn broadcast() -> Self {
-        ShortAddress(0xffff)
-    }
-
-    /// Decodes an address from a byte buffer
-    ///
-    /// This method is used by [`Header::decode`] to decode addresses. Unless
-    /// you decide to write your own code for decoding headers, there should be
-    /// no reason to call this method directly.
-    ///
-    /// # Errors
-    ///
-    /// This function returns an error, if there are not enough bytes in the
-    /// buffer to encode a valid `Address` instance.
-    ///
-    /// # Example
-    ///
-    /// ``` rust
-    /// use ieee802154::mac::ShortAddress;
-    ///
-    /// # fn main() -> Result<(), ::ieee802154::mac::DecodeError> {
-    /// let mut bytes = &[0x56, 0x78][..];
-    /// let address = ShortAddress::decode(&mut bytes)?;
-    ///
-    /// assert_eq!(address, ShortAddress(0x7856));
-    /// #
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn decode(buf: &mut dyn Buf) -> Result<Self, DecodeError> {
-        if buf.remaining() < 2 {
-            return Err(DecodeError::NotEnoughBytes);
-        }
-
-        Ok(ShortAddress(buf.get_u16_le()))
-    }
-
-    /// Encodes the address into a buffer
-    ///
-    /// # Example
-    ///
-    /// ``` rust
-    /// use ieee802154::mac::ShortAddress;
-    ///
-    /// let address = ShortAddress(0x5678);
-    ///
-    /// let mut bytes = bytes::BytesMut::with_capacity(2);
-    /// address.encode(&mut bytes);
-    ///
-    /// let expected_bytes = [0x78, 0x56];
-    /// assert_eq!(bytes[..], expected_bytes[..]);
-    /// ```
-    pub fn encode(&self, buf: &mut dyn BufMut) {
-        buf.put_u16_le(self.0)
-    }
-}
-
-/// A 64-bit extended address
-///
-/// A unique address that is used to identify an device in the PAN.
-///
-/// # Example
-///
-/// ``` rust
-/// use ieee802154::mac::frame::ExtendedAddress;
-///
-/// let ext_address = ExtendedAddress(0x0123456789abcdef);
-/// ```
-#[derive(Clone, Copy, Debug, Eq, Hash, Hash32, PartialEq)]
-pub struct ExtendedAddress(pub u64);
-
-impl ExtendedAddress {
-    /// An instance of `ExtendedAddress` that represents the broadcast address.
-    pub const BROADCAST: Self = ExtendedAddress(0xffffffffffffffffu64);
-
-    /// Creates an instance of `ExtendedAddress` that represents the broadcast address
-    pub fn broadcast() -> Self {
-        ExtendedAddress(0xffffffffffffffffu64)
-    }
-
-    /// Decodes an address from a byte buffer
-    pub fn decode(buf: &mut dyn Buf) -> Result<Self, DecodeError> {
-        if buf.remaining() < 8 {
-            return Err(DecodeError::NotEnoughBytes);
-        }
-        Ok(ExtendedAddress(buf.get_u64_le()))
-    }
-
-    /// Encodes the address into a buffer
-    pub fn encode(&self, buf: &mut dyn BufMut) {
-        buf.put_u64_le(self.0)
-    }
-}
-
-/// An address that might contain an PAN ID and address
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub enum Address {
-    /// Short (16-bit) address and PAN ID (16-bit)
-    Short(PanId, ShortAddress),
-    /// Extended (64-bit) address and PAN ID (16-bit)
-    Extended(PanId, ExtendedAddress),
-}
-
-impl Address {
-    /// Creates an instance of `Address` that represents the broadcast address
-    pub fn broadcast(mode: &AddressMode) -> Option<Self> {
-        match mode {
-            AddressMode::None => None,
-            AddressMode::Short => Some(Address::Short(
-                PanId::broadcast(),
-                ShortAddress::broadcast(),
-            )),
-            AddressMode::Extended => Some(Address::Extended(
-                PanId::broadcast(),
-                ExtendedAddress::broadcast(),
-            )),
-        }
-    }
-
-    /// Decodes an address from a byte buffer
-    pub fn decode(buf: &mut dyn Buf, mode: &AddressMode) -> Result<Option<Self>, DecodeError> {
-        let opt_address = match mode {
-            AddressMode::None => None,
-            AddressMode::Short => {
-                let pan_id = PanId::decode(buf)?;
-                let short = ShortAddress::decode(buf)?;
-                Some(Address::Short(pan_id, short))
-            }
-            AddressMode::Extended => {
-                let pan_id = PanId::decode(buf)?;
-                let extended = ExtendedAddress::decode(buf)?;
-                Some(Address::Extended(pan_id, extended))
-            }
-        };
-        Ok(opt_address)
-    }
-
-    /// Decodes an address from a byte buffer
-    pub fn decode_compress(
-        buf: &mut dyn Buf,
-        mode: &AddressMode,
-        pan_id: PanId,
-    ) -> Result<Option<Self>, DecodeError> {
-        let opt_address = match mode {
-            AddressMode::None => None,
-            AddressMode::Short => {
-                let short = ShortAddress::decode(buf)?;
-                Some(Address::Short(pan_id, short))
-            }
-            AddressMode::Extended => {
-                let extended = ExtendedAddress::decode(buf)?;
-                Some(Address::Extended(pan_id, extended))
-            }
-        };
-        Ok(opt_address)
-    }
-
-    /// Encodes the address into a buffer
-    pub fn encode(&self, buf: &mut dyn BufMut) {
-        match *self {
-            Address::Short(pan_id, short) => {
-                pan_id.encode(buf);
-                short.encode(buf);
-            }
-            Address::Extended(pan_id, extended) => {
-                pan_id.encode(buf);
-                extended.encode(buf);
-            }
-        }
-    }
-
-    /// Encodes the address into a buffer
-    pub fn encode_compress(&self, buf: &mut dyn BufMut) {
-        match *self {
-            Address::Short(_, a) => a.encode(buf),
-            Address::Extended(_, a) => a.encode(buf),
-        }
-    }
-
-    /// Get the PAN ID for this address
-    pub fn pan_id(&self) -> PanId {
-        match *self {
-            Address::Short(pan_id, _) => pan_id,
-            Address::Extended(pan_id, _) => pan_id,
-        }
-    }
-
-    /// Get the address mode for this address
-    pub fn address_mode(&self) -> AddressMode {
-        match *self {
-            Address::Short(_, _) => AddressMode::Short,
-            Address::Extended(_, _) => AddressMode::Extended,
-        }
-    }
-}
-
 /// Content of a frame
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum FrameContent {
@@ -717,7 +244,7 @@ pub enum FrameContent {
 impl FrameContent {
     /// Decode frame content from byte buffer
     pub fn decode(buf: &mut dyn Buf, header: &Header) -> Result<Self, DecodeError> {
-        match header.frame_control.frame_type {
+        match header.frame_type {
             FrameType::Beacon => {
                 let beacon = Beacon::decode(buf)?;
                 Ok(FrameContent::Beacon(beacon))
@@ -767,7 +294,9 @@ mod tests {
     use super::*;
     use crate::mac::beacon;
     use crate::mac::command;
-    use crate::mac::FrameVersion;
+    use crate::mac::header::{
+        Address, ExtendedAddress, FrameVersion, PanId, Security, ShortAddress,
+    };
     use bytes::BytesMut;
 
     #[test]
@@ -777,13 +306,13 @@ mod tests {
         ];
         let mut sliced_data = &data[..];
         let frame = Frame::decode(&mut sliced_data, true).unwrap();
-        let fctl = frame.header.frame_control;
-        assert_eq!(fctl.frame_type, FrameType::Data);
-        assert_eq!(fctl.security, false);
-        assert_eq!(fctl.frame_pending, false);
-        assert_eq!(fctl.ack_request, false);
-        assert_eq!(fctl.pan_id_compress, true);
-        assert_eq!(fctl.version, FrameVersion::Ieee802154_2003);
+        let hdr = frame.header;
+        assert_eq!(hdr.frame_type, FrameType::Data);
+        assert_eq!(hdr.security, Security::None);
+        assert_eq!(hdr.frame_pending, false);
+        assert_eq!(hdr.ack_request, false);
+        assert_eq!(hdr.pan_id_compress, true);
+        assert_eq!(hdr.version, FrameVersion::Ieee802154_2003);
         assert_eq!(
             frame.header.destination,
             Some(Address::Short(PanId(0x208f), ShortAddress(0xffff)))
@@ -815,13 +344,13 @@ mod tests {
             0x4a, 0xc2, 0xae, 0xaa, 0xbb, 0xcc,
         ];
         let frame = Frame::decode(&data[..], true).unwrap();
-        let fctl = frame.header.frame_control;
-        assert_eq!(fctl.frame_type, FrameType::Data);
-        assert_eq!(fctl.security, false);
-        assert_eq!(fctl.frame_pending, false);
-        assert_eq!(fctl.ack_request, true);
-        assert_eq!(fctl.pan_id_compress, false);
-        assert_eq!(fctl.version, FrameVersion::Ieee802154_2003);
+        let hdr = frame.header;
+        assert_eq!(hdr.frame_type, FrameType::Data);
+        assert_eq!(hdr.security, Security::None);
+        assert_eq!(hdr.frame_pending, false);
+        assert_eq!(hdr.ack_request, true);
+        assert_eq!(hdr.pan_id_compress, false);
+        assert_eq!(hdr.version, FrameVersion::Ieee802154_2003);
         assert_eq!(
             frame.header.destination,
             Some(Address::Short(PanId(0xffff), ShortAddress(0x0002)))
@@ -840,16 +369,12 @@ mod tests {
     fn encode_ver0_short() {
         let frame = Frame {
             header: Header {
-                frame_control: FrameControl {
-                    frame_type: FrameType::Data,
-                    security: false,
-                    frame_pending: false,
-                    ack_request: false,
-                    pan_id_compress: false,
-                    version: FrameVersion::Ieee802154_2003,
-                    dest_addr_mode: AddressMode::Short,
-                    src_addr_mode: AddressMode::Short,
-                },
+                frame_type: FrameType::Data,
+                security: Security::None,
+                frame_pending: false,
+                ack_request: false,
+                pan_id_compress: false,
+                version: FrameVersion::Ieee802154_2003,
                 destination: Some(Address::Short(PanId(0x1234), ShortAddress(0x5678))),
                 source: Some(Address::Short(PanId(0x4321), ShortAddress(0x9abc))),
                 seq: 0x01,
@@ -874,16 +399,12 @@ mod tests {
     fn encode_ver1_extended() {
         let frame = Frame {
             header: Header {
-                frame_control: FrameControl {
-                    frame_type: FrameType::Beacon,
-                    security: false,
-                    frame_pending: true,
-                    ack_request: false,
-                    pan_id_compress: false,
-                    dest_addr_mode: AddressMode::Extended,
-                    version: FrameVersion::Ieee802154_2006,
-                    src_addr_mode: AddressMode::Short,
-                },
+                frame_type: FrameType::Beacon,
+                security: Security::None,
+                frame_pending: true,
+                ack_request: false,
+                pan_id_compress: false,
+                version: FrameVersion::Ieee802154_2006,
                 destination: Some(Address::Extended(
                     PanId(0x1234),
                     ExtendedAddress(0x1122334455667788),
@@ -923,16 +444,12 @@ mod tests {
     fn encode_ver0_pan_compress() {
         let frame = Frame {
             header: Header {
-                frame_control: FrameControl {
-                    frame_type: FrameType::Acknowledgement,
-                    security: false,
-                    frame_pending: false,
-                    ack_request: false,
-                    pan_id_compress: true,
-                    dest_addr_mode: AddressMode::Extended,
-                    version: FrameVersion::Ieee802154_2003,
-                    src_addr_mode: AddressMode::Short,
-                },
+                frame_type: FrameType::Acknowledgement,
+                security: Security::None,
+                frame_pending: false,
+                ack_request: false,
+                pan_id_compress: true,
+                version: FrameVersion::Ieee802154_2003,
                 destination: Some(Address::Extended(
                     PanId(0x1234),
                     ExtendedAddress(0x1122334455667788),
@@ -961,16 +478,12 @@ mod tests {
     fn encode_ver2_none() {
         let frame = Frame {
             header: Header {
-                frame_control: FrameControl {
-                    frame_type: FrameType::MacCommand,
-                    security: false,
-                    frame_pending: false,
-                    ack_request: true,
-                    pan_id_compress: false,
-                    dest_addr_mode: AddressMode::None,
-                    version: FrameVersion::Ieee802154,
-                    src_addr_mode: AddressMode::Short,
-                },
+                frame_type: FrameType::MacCommand,
+                security: Security::None,
+                frame_pending: false,
+                ack_request: true,
+                pan_id_compress: false,
+                version: FrameVersion::Ieee802154,
                 destination: None,
                 source: Some(Address::Short(PanId(0x1234), ShortAddress(0x9abc))),
                 seq: 0xff,
