+++ src/mac/mod.rs
@@ -4,9 +4,4 @@ pub mod beacon;
 pub mod command;
 pub mod frame;
 
-pub use frame::{
-    Address, DecodeError, ExtendedAddress, Frame, FrameContent, Header, PanId, ShortAddress,
-    WriteFooter,
-};
-
-pub use frame::frame_control::{AddressMode, FrameControl, FrameType, FrameVersion};
+pub use frame::header;
